// packages/web/src/gen/openapi.ts
export interface components {
    schemas: {
        Coord: {
            /** Format: double */
            lat: number;
            /** Format: double */
            lng: number;
        };
        Location: {
            name: string;
            address: string;
            coord: components["schemas"]["Coord"];
        };
        RouteEstimate: {
            from: components["schemas"]["Location"];
            to: components["schemas"]["Location"];
            /** Format: int32 */
            distanceMeters: number;
            /** Format: int32 */
            durationSeconds: number;
        };
        MeetingOption: {
            destination: components["schemas"]["Location"];
            fromA: components["schemas"]["RouteEstimate"];
            fromB: components["schemas"]["RouteEstimate"];
            /** Format: int32 */
            totalDurationSeconds: number;
            /** Format: int32 */
            durationDifferenceSeconds: number;
            /** Format: int32 */
            maxDurationSeconds: number;
        };
        CompareMeetingRequest: {
            locationA: components["schemas"]["Location"];
            locationB: components["schemas"]["Location"];
            interestAreas: components["schemas"]["Location"][];
        };
        CompareMeetingResponse: components["schemas"]["MeetingOption"][];
        ErrorResponse: {
            error: string;
            code: string;
            details: string;
        };
        HealthResponse: {
            ok: boolean;
            message: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}